+++ ui/components/multichain/disconnect-all-gator-permissions-modal/disconnect-all-gator-permissions-modal.tsx
