+++ ui/components/multichain/disconnect-all-gator-permissions-modal/index.ts
@@ -1,4 +1,4 @@
-export { DisconnectAllGatorPermissionsModal } from './disconnect-permissions-modal';
+export { DisconnectAllGatorPermissionsModal } from './disconnect-all-gator-permissions-modal';
 export { PermissionItem } from './permission-item';
 export type {
   DisconnectAllGatorPermissionsModalProps,
