+++ ui/components/multichain/disconnect-all-gator-permissions-modal/permission-item.tsx
