+++ ui/components/multichain/disconnect-all-gator-permissions-modal/types.ts
